Update Program.cs
Volume controls with arrow up and down are working but you have to wait for the next song to play but you can skip the song after updating volume and that work. Need to make a function to handle updating the volume as we don't have any error checking on the value right now. Do that next.